Includes the following:
(1) New function: eval_ignore_buck_rules('<path>') This evaluates the
build rules in <path> within the current globals, except that it
suppresses (ignores) calls to any regular BUCK rules like java_library.
The net impact is to turn the BUCK file into a metadata store since it
no longer contributes rules.

(2) New function: is_build_file_context(). This returns True only if the
current context is a toplevel build file parsing. That is, if this
function returns true, it is a signal that BUCK file is being parsed at
the top level and if it returns false, it is a signal that the current
evaluation is a metadata evaluation.